backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
pandas==2.2.2
openai>=1.50.0
httpx==0.27.2
python-multipart==0.0.9
requests==2.31.0
# --- FIX DEFINITIVO DE DEPENDENCIAS CRIPTOGRÁFICAS ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
PyNaCl==1.5.0
# ----------------------------------------------------